+++ bedrock/bedrock-v465/src/main/java/com/nukkitx/protocol/bedrock/v465/serializer/AddVolumeEntitySerializer_v465.java
@@ -1,4 +1,4 @@
-package com.nukkitx.protocol.bedrock.v448.serializer;
+package com.nukkitx.protocol.bedrock.v465.serializer;
 
 import com.nukkitx.protocol.bedrock.BedrockPacketHelper;
 import com.nukkitx.protocol.bedrock.packet.AddVolumeEntityPacket;
